feat(sources): Add Sqs source
Added the first draft of a Sqs queue as a source for a pipe.
It uses the properties provided by the generated cfn so that they are
automatically updated together with cfn changes.

The PipeSource had to be updated so that the sourceArn can be set by the
implementing class.